core/merge: Save local file metadata separately (#1904)
We want to keep track of the current local state of files, whether
they're in sync or not.

At the moment, we store altogether in our PouchDB records the 3
different states of documents — local, remote and last synced — and a
few technical attributes that help us make sense of it all (e.g. which
side has received changes that need to be synchronized, what actions
need to be applied…).
This mechanism is limited because we don't store separately the values
of attributes on each side. For example, if a file's `size` has
changed on the remote Cozy and we've saved it in PouchDB, we've lost
the local file's `size`.
When loosing this information, it's harder for us to make sense of our
data and to make the right decision and even impossible to make some.

A good example of such an impossible decision to make can be described
by this scenario:
0. A file `a` is synchronized between the local filesystem and the
   remote Cozy
1. `a` is updated on the remote Cozy and thus its `md5sum` stored in
   PouchDB is changed
2. the client is stopped before the change is applied on the
   local filesystem

⇒ When the client is restarted, the local scan will generate an event
with the local, outdated, file `md5sum` which will be treated as an
update because we see a difference between the event's `md5sum` and
the one stored in PouchDB. We have no way to tell this was the
previous `md5sum` and this change should be discarded.

By storing the local state separately, we have this information and we
can use it to tell whether the detected local change is an actual one
or not by comparing the stored local state and the new one.

This commit only introduces local file states, the methods to update
them, to compare them with one another and the calls to keep them
up-to-date with the actual file's local state.