core/merge: Discard same-local-state files changes (#1905)
We'll stop merging local file updates if the local metadata stored in
PouchDB represent the same file as the metadata of the update.

When applying remote file changes to the local filesystem, the local
watcher picks up events that can lead to a local update request with
outdated metadata (i.e. the previous version of the metadata).

To avoid merging those fake updates and propagate inaccurate changes
to the remote Cozy, we'll use the newly introduced `local` metadata to
check whether the update is a true one or it corresponds to the
previous version of the file being synchronized.

This should prevent the creation of conflicts when propagating the
combined renaming and update of a Cozy Note to the local filesystem.