local/chokidar/watcher: Normalize changes paths (#1910)
We'll correct paths normalization on local changes instead of local
events (i.e. after the analysis step) as well as improve the
correction by correcting paths of new documents when they're added
into an existing folder.

On macOS, saved paths in PouchDB records and actual paths on the
filesystem can differ by their UTF-8 normalization. Indeed, documents
created on the remote Cozy will always be normalized with NFC while
documents created locally **can** be normalized with NFD. This is the
case on HFS+ volumes for example.
The local normalization is done automatically and generates changes
that we don't want to propagate to the remote Cozy. This is why the
normalization can be different between PouchDB and the filesystem.

We'll consider that the local filesystem is applying an NFD
normalization to document names.

When a document is modified locally, the event's path will be
normalized with NFD. We have some logic to match this document with
its PouchDB record and "correct" the normalization if it was first
saved with an NFC encoded path.

However, this correction would only be applied if a document was
already saved in PouchDB and on local events.
The first restriction means that documents added in an existing folder
won't get their path corrected which can lead to file paths completely
encoded with NFD while their parent path is encoded with NFC in
PouchDB for example.
The second restriction means we can't correct moves and renamings
paths since those are made of two separate events (i.e. `unlink` and
`add` or `unlinkDir` and `addDir`).

Now that the `analysis` step compares event paths in their normalized
form, we can move the normalization correction into a new step coming
later and applied on local changes like `FileAddition` and `FileMove`.