local/atom/incomplete_fixer: File update via backup (#1912)
Some software take a lot of precautions when saving modifications made
on a file to avoid any data loss.
One such software moves the current file to a backup location and
writes the new version in its stead.

In this situation, on Windows, the Atom watcher receives and processes
the following events:
1. `renamed` (`file` → `file.backup`)
2. `created` (`file`)
3. `deleted` (`file.backup`)

If the backup file deletion was done before the watcher could compute
its checksum, the `renamed` event will be marked as incomplete. It
will then be completed in the `incompleteFixer` step by the `deleted`
event, transforming the `renamed` event into a `deleted` event of the
original file location (i.e. the same as the created file location
since we're replacing it).

This would be alright if the `created` event was fired and thus
processed after the `deleted` event. We would then group those two
events together to form an overwriting `created` event.
However, in this case, the `created` event is processed before and
therefore the `deleted` event will be merged last.

We would then propagate a file deletion to the remote Cozy when the
file still exists on the local filesystem.

We propose here to complete the incomplete `renamed` event with the
`created` event. They can be linked together via the path (i.e.
`oldpath` for the renamed event and `path` for the created one).
The completed event would be a `modified` event with the stats and
checksum of the `created` event.

The remaining `deleted` event for the backup file will have no effect
since we shouldn't have any record for it in PouchDB. If we did have
one then removing it would be the right thing to do.