core/merge: Compare only existing local states (#1913)
We have recently introduced a local state for files via a `local`
attribute in their PouchDB records.

When the file is created on the local filesystem,  the local state
will be filled when the record is created.
On the other hand, when the file is created on the remote Cozy, the
state won't exist until the local watcher detects the file's creation
on the local filesystem and the generated `created` event is merged.

On Windows, this `created` event can be ignored if we're replacing an
existing document. We'd have a prior `deleted` event at the same
path and the two events would be grouped together by the
`winDetectMove` step into an `ignored` event. We don't have a
`renamed` event because the source and destination paths would be the
same and the result move would be invalid.
In this case, only the `modified` events following a newly downloaded
file creation (i.e. for the executable permission and dates changes)
would be dispatched.

In case of a file update, we were expecting the local state to already
exist but in our situation, it does not.
From there, when we try to compare the existing record's local state
(i.e. the missing one) and the newly built local state the
`Cannot read property "executable" of undefined` error is raised.

To avoid raising this error, we can simply check whether or not the
existing record has a local state before trying to compare it with the
new state.