core/merge: Move children with different normalization (#1918)
When moving a folder whose path is normalized with a certain norm
(i.e. NFC or NFD) but its children's paths are normalized with another
one (at least the parent part), we want to make sure the new
children's paths are correctly computed (i.e. the parent part is
replaced with its new path) and that their id is in sync with their
new path.

When a folder is moved, we fetch all its descendants from PouchDB via
their ids, based on the folder's id (i.e. we look for every document
whose id starts with the folder id). In this case, the path
normalization doesn't matter because ids are re-normalized.
Once we have all the descendants records, we replace the folder part
in its descendants' paths with its new path. However, if the
normalization is different, the replacement won't work (the String
method won't find the substring to replace).

e.g.:
- We have the folder `énoncés/` formatted with NFC
- We have the file `énoncés/DS-1.pdf` where énoncés is formatted with
  NFD
When renaming `énoncés/` to `corrigés/`, the file's path won't be
changed while its id will be based on `corrigés/DS-1.pdf`.

Not changing the path when the id has changed can lead to a lot of
errors later since we base most of our logic to detect movements on
the path but access to the PouchDB records is done via the id.

We introduce the method `metadata.newChildPath()` which will replace
the old parent path with the new in the child path no matter what the
normalizations are of the moved folder and its ancestors.
This way we make sure the children are correctly moved with their
parent and their id is kept in sync with their path.