core/utils/timestamp: Round non-UTC remote dates (#1922)
We should correctly round dates sent by the remote Cozy whether
they're in UTC format or not.

We recently started to round the `created_at` and `updated_at` dates
coming from the remote Cozy to make sure we have the expected ISO
format in our PouchDB records.
However, we were expecting to receive dates in UTC only while we can
receive dates in other timezones.
When the timezone is not UTC, our rounding function raises an error
after trying to read the `length` of `undefined`.

This error is raised because our method to get the milliseconds part
of the date as a string relied on the presence of the letter `Z` at
the end of the date string.
When the timezone is not UTC, the string ends with the time offset and
the milliseconds part could be `undefined` if milliseconds were not
present in the date.

We propose a new method based on a regular expression looking for the
presence of a dot followed by digits. If those are present then we
extract the digits as the milliseconds string. If they're not present,
we assign an empty string to the milliseconds.
In both cases, we can safely look at the `length` of the milliseconds.