gui: Don't automatically re-enable autolaunch (#1938)
We introduced a fix a while back to make sure the autolaunch
`.desktop` file created on Linux would use the app's name and not its
path since the path would contain the version by default and
autolaunch would thus break after each upgrade.

We tried to re-enable autolaunch with the app's name for users who had
it enabled with the app's path. We would check the return value of
`autoLauncher.isEnabled()` to decided whether we needed to do it but
this value is not a boolean as expected, it is a promise of a boolean.
This means that we would always try to re-enable autolaunch, at each
run, for users who wanted to disable it.

We're now waiting for the promise resolution and use the resolved
value to decide if we need to re-enable autolaunch.