test/builders/metadata: Use valid remote docs (#1966)
The test data builders generating `Metadata` objects need to populate
the `remote` attribute of those objects when building records of
documents which have been synchronized with the remote Cozy (or are
coming from the Cozy).

However, those builders were still only storing the remote id and
revision of documents instead of the full document.

We now use the remote data builders to generate remote documents of
the appropriate type (i.e. files or directories) when the record being
built is supposed to have a `remote` attribute.