test/integration/notes: Update time with note (#1968)
When updating a remote document in tests, we should always set the
`updated_at` attribute to the update time or the local `Metadata`
object will keep the latest update time and we'll see synchronization
issues arise once the local watcher kicks in (i.e. the update time,
which is used to set the local file metadata on the filesystem, hasn't
changed so we won't recompute the checksum of the local file although
its content has changed).

For now the `RemoteBaseBuilder.update()` method is only used once in
the Notes integration tests.

We also fix the `timestamp.spread()` method which was returning a
0-based month number when builders and `timestamp` methods are
expecting a 1-based month when building dates.