core/remote: End rejections handling with request (#1969)
We introduced unhandled promise rejections "handling" via the
`process` `unhandledRejection` event listener to make sure upload
requests will always terminate even when the underlying chromium
request fails with an unhandled promise rejection.

However, if the request succeeds, we don't remove the event listener
and thus we'll keep adding listeners with every upload.
We now remove the listener even if it was not triggered (i.e. the use
of `once` would remove it in this case).

This avoids `MaxListenersExceededWarnings` especially during tests.