core/metadata: Require `executable` file attribute (#1976)
The `executable` attribute of `Metadata` objects is marked as optional
in the type definition. This leads to uncertainty as to whether the
attribute is defined in `Metadata` objects or not and forces us to
check for its presence every time we want to read or set it. Besides,
it allows us to build file `Metadata` objects without it.

Marking the attribute as required and setting it for every file
`Metadata` object will help us ensure our documents are valid and
simplify the use of this attribute.

/!\ Although it is marked as required, it will still be missing in
folder `Metadata` objects. This is made possible by cracks in our
types coverage but should be fixed once we'll have different type
definitions for file and folder metadata.