core/sync: Compute current revision in applyDoc (#1977)
Instead of computing the current document "revision" when computing
the out-of-date side and passing it down to `applyDoc()` as an
argument as well as the side name, we can compute it within
`applyDoc()` where it's used.

This will simplify `applyDoc()` unit tests as we won't have to
determine what's the appropriate `rev` (i.e. the previous term) to
pass to `applyDoc()` to have the expected behavior.