core/metadata: Mark `tags` attribute as required (#1981)
The `Metadata` `tags` attribute could either be missing or a list of
string values. It is now marked as required.

The `metadata.buildFile()` and `metadata.buildFolder()` methods will
now always default it to an empty Array since tags come from remote
documents and those are used to build `Metadata` objects from local
data.

The `merge` and `move` methods won't need to default the attribute to
an empty Array anymore and moved documents will inherit the list of
tags from the source document.