core/merge: Fix conflicting file updates (#1986)
To avoid losing merged but unsynced remote file updates while avoiding
unnecessary conflicts, we introduced an exception in the
`Merge.updateFileAsync()` behavior when merging an "update" from an
initial scan event.

In this case, we would always consider that the file had not been
updated on the local file system to preserve the remote update.
It means that if the file had actually been updated, we would lose the
local update.
To mitigate this problem, we would not mark the update as an
overwriting one, thus telling Sync to trash a copy of the file before
applying the remote update on the local filesystem.

Now that we track the local metadata separately from the remote and
"merged" metadata, we can tell if the file was updated locally or not.
If it wasn't we won't modify the existing PouchDB record with the
unsynced remote update. This update will then be applied on the local
filesystem as an overwriting one, thus without any trashed copy of the
local version.
If it was updated, the local update will be considered as a
conflicting one and we'll treat it as such by dissociating the local
and remote versions and renaming the local file with a conflict
suffix.