core/merge: Merge attributes in updateFile (#1987)
When we merge a file update in PouchDB, we fetch the existing record
from PouchDB and copy some of its attributes to the new record for the
updated file.
However, we were not copying all relevant attributes and thus losing
some of them (e.g. the `tags` attribute would be overwritten).

To avoid losing data and make this copy more scalable, we're now
building the new record based on the existing one and overwriting
only the relevant attributes.

This process should probably be done in every `Merge` method updating
an existing PouchDB record.