gui: Fix syncing status display (#2004)
The syncing status comes from the Sync state and is propagated to the
main GUI window and the systray icon. It can come with data (i.e.
metadata about the file being transferred) from which we want to
derive a label for the systray.

We had 2 issues with this process:
1. the filename was not passed correctly to the `tray` module in
   charge of setting the systray label
2. the method in charge of setting the systray label would not
   correctly concatenate the main part of the label with the given
   filename

We also take the opportunity to harmonize the status name across the
board by using `syncing` instead of `sync` in the Sync state.