core/local: Detect ACL lock on open files (#2011)
It appears that the Node `fs.access()` method does not check
access-control policies on files and folders.
This means that `fs.access()` won't detect when a file is opened in
an editor like Office and is thus locked (and its parent folders as
well).

As a workaround, we can use the `fs.open()` method with a read/write
access flag. If the file is locked, the call will fail with a `EBUSY`
error.
However, this won't detect locked folders and our only option seems to
try applying changes on the folder and manage errors.