core/app: Unregister client even with deleted Cozy (#2020)
When the user tries to unregister their Desktop client from its remote
Cozy, we attempt to:
1. delete the OAuth client on the remote Cozy
2. and then delete the local client config

If the remote Cozy does not exist anymore because it's either been
deleted on moved to another location (and 30 days have passed since),
we won't be able to delete the OAuth client.
In this situation we still want to delete the local config so the user
can connect its Desktop client to another Cozy (e.g. the new
location).