core/remote: Use most recent modification date (#2023)
When sending a new file to a remote Cozy, it can happen that the stack
saves in CouchDB a creation date more recent than the one provided in
the upload request.

For example. if the uploaded file is:
- a photo with EXIF metadata
- and the creation date in the metadata does not contain a timezone
- and the photo was taken in a timezone ahead of UTC

In this case, the stack will use the EXIF date as creation date and
consider it as a UTC datetime. Therefore, the creation date as saved
on the remote Cozy will be ahead of the creation date on the local
filesystem.

If this photo is then moved on the local filesystem for example, the
stack would refuse the move as the provided modification date (i.e.
the one taken from the filesystem) will be older than the remote one.
This leads to desynchronization between the computer and the Cozy and
thus potential for more problems in the future.

As a simple mitigation, we now always use the most recent modification
date between the local and the remote one when sending actions to the
remote Cozy.
The next step would be to propagate this new date to the local
filesystem.