gui/js/lastfiles: Persist same path only once (#2026)
The list of recently synchronized files is persisted to disk in a JSON
file. We only keep the last 250 files to keep it short.

However, it happens that we would not remove existing occurrences of
the same path when adding a file to the list although we only display
the latest occurrence in the GUI. This means that if we were to
synchronize 250 modifications on the same file. we would only display
this one file in the GUI while the list on disk would be full.

We now make sure only one occurrence is saved on disk.