core/sync: Restart remote watcher when back online (#2027)
When the remote watcher fails to fetch remote changes because the Cozy
is unreachable (i.e. we received a FetchError with a status for which
we don't have a special treatment or that treatment is notifying that
the Cozy is unreachable) we block the synchronization and stop the
remote watcher to prevent further errors.

While we were regularly pinging the remote Cozy to see if it was
reachable again and notifying the GUI with an `online` event in this
case, we were not restarting the remote watcher.
Therefore, we were not fetching remote changes anymore after the Cozy
was made unreachable once.

We now make sure the remote watcher is restarted whenever we unblock
the synchronization.