core/remote: Remove _type attribute from doc types (#2032)
The `_type` attribute is used by `cozy-stack` to store the doctype of
CouchDB records. In our case, this doctype is always `io.cozy.files`
and since it is not stored in the CouchDB records themselves, the data
we receive from our calls to the remote changes feed for example does
not contain this attribute.

We don't need this attribute and should not expect its presence in
returned remote records via our Flow types nor add it ourselves since
it makes it harder to compare remote data in tests with docs fetched
from our helpers.