test/builders/metadata: Keep up-to-date sides (#2036)
Our Metadata builders allow us to generate accurate Metadata objects
from scratch or built either from a remote document or another
Metadata object or even both.
However, to accommodate for a lot of different test contexts, the logic
for ensuring we have valid `local` and `remote` attributes is quite
complex and we missed a spot for built Metadata objects marked as
up-to-date (i.e. their `sides` attribute's `local` and `remote` values
are both equal to the `target` value which means the represented
document is synced).

To cover this case, we introduce 2 functions that will help us
determine if the `local` and `remote` attributes need to be updated or
should remain untouched.
These methods are meant to be modified in case new attributes should
be considered to make the decision.