core/metadata: Round updated remote date times (#2037)
When a creation or modification date is generated by the remote Cozy,
it has a nanosecond precision, formatted as extra millisecond digits.

When those dates are converted by the Desktop client (for comparison
purposes) via the `Date` constructor, those extra digits are simply
dropped. The resulting date is thus older than the original one.

This can prevent future requests to the Cozy to complete so we
introduced a method to round those dates up to the next millisecond
instead.
However, we used it only during the creation of documents on the
remote Cozy and forgot to use it when dealing with updated remote
documents.