core/remote: Don't sanitize doc names in requests (#2041)
By default `cozy-client-js` will sanitize the names of documents (i.e.
trim them) when sending requests to the remote Cozy.
This is a problem for the Desktop client as we won't update local
names if their remote counterparts were sanitized and we can end up
with missing parent documents or other similar errors.

Since v0.18.0, `cozy-client-js` has a `noSanitize` option for all
requests that would sanitize names by default. If this option is
passed and set to `true`, `cozy-client-js` will refrain from
sanitizing.