core/merge: Stop creating local parent records (#2058)
When merging a local change, we were checking if its parent folder
record existed in PouchDB. To make sure the whole hierarchy was
present in the database, we would create any missing parent folder
record with only a few attributes: their path and update date.

However, this generates records missing most attributes that we rely
on more and more (i.e. inode, local and remote attributes…) and should
not be necessary as a document cannot exist on the local filesystem
without its parent folder and so we should receive an event leading us
to merge that parent folder record at some point.

This is why we'll be more optimistic and rely on an eventual
consistency regarding the parent hierarchy of any given record in
PouchDB.
Besides, we were not creating those parent records for remote
documents anymore.