core/sync: Update errors only when about to retry (#2060)
The Sync retry mechanism includes incrementing the failed change
errors counter in its PouchDB record. This step is also useful in some
cases where we made some changes to the record when catching the
propagation error (e.g. when failing to update a remote file after
successfully moving it) and need to save them to PouchDB to make sure
they won't be picked up later (when retrying the propagation for
example).

However, saving the record will generate a new PouchDB sequence for
the change and will prevent the user from manually skipping it as this
relies on marking the change's sequence as processed. In this
situation, we'll still retry the failed change after the user has
requested to skip it.

To allow for both behaviors to coexist, we'll only save the modified
record with its incremented errors counter right before starting a new
Sync run (i.e. retrying) as the user will have had the opportunity to
skip the change by that time and can't while Sync is running.

This does not address any issues that could arise from the client
stopping before we get to save the record modifications as those
already exist although they might be harder to encounter as we save
the record sooner.