ci: Cancel outdated Github Workflow runs (#2064)
CI resources on Github are sparse especially for macOS jobs. This
means that running a lot of macOS jobs can increase their queuing time
and thus the time we have to wait to know if tests are green.

To avoid losing too much time, we will now cancel all workflow runs
but the latest one for each branch on push.
We do so by using the `styfle/cancel-workflow-action` action, with a
specific commit that is not merged yet but offers this feature.