core/merge: Fix trashing of overwriting moves (#2072)
When we took care of trashing the overwritten document of a trashed
overwriting move, we made the wrong assumption that the overwritten
and overwriting records would not have the same PouchDB `_id` and thus
could be modified together via a `Pouch.bulkDocs()` call.

It turns out that we reuse the _id (and _rev) of overwritten records
when merging an overwriting move so that only one record with the
destination path exists in PouchDB at the same time (i.e. to avoid
queries mismatches).
Thus, when trashing an overwriting move we'll only update the
destination record as if it was the deleted overwritten record.

We did not catch this problem because our test Metadata builders did
not correctly emulate an overwriting move as they were not reusing the
overwritten record's _id and _rev attributes.