core/sync: Prevent parallel retry intervals (#2080)
When the remote Cozy becomes unreachable, two processes can throw
errors which will block the synchronization process and initiate a
retry interval:
- the synchronization process itself if the error was thrown while
  trying to propagate a change
- the remote watcher when trying to fetch the latest remote changes

In the synchronization process errors out right before the remote
watcher tries to fetch the latest changes (i.e. the Cozy is thus still
unreachable), the reference to the retry interval initiated by the
synchronization process will be overwritten by the reference to the
retry interval initiated by the remote watcher.
This reference is held by the `Sync.retryInterval` variable.

Losing the reference to the 1st retry interval means it will run in
parallel to the 2nd one and we won't be able to end it via a call to
`clearInterval()`.

To prevent this situation from happening, we'll clear any existing
interval held by `Sync.retryInterval` before starting a new one and
assigning it to the same variable.

---

In order to limit unnecessary computations and extra errors, we'll
also stop the remote watcher whenever the synchronization process is
blocked because the remote Cozy is unreachable.
It will be restarted by the retry mechanism when the Cozy becomes
available again.