gui: Avoid unwanted note openings (#2081)
When the client is launched with more than 2 arguments (the first 2
being the path to the application and an empty string), we assume the
last one is the path to a Cozy Note on the file system as this is the
way Linux and Windows notify us of an opening request from the file
explorer.

However, when the client is started automatically with the computer,
the `--hidden` argument is added by `auto-launch` to make sure the
client's main window is not opened right away.
In this case, the client will assume `--hidden` is the path to a note
and will try to open it. Since it's not a path at all, it will fail
with a generic error displayed in a popup.

This error does not prevent the client from starting however.

We'll now check if the "path" ends with `.cozy-note` to filter out all
situations where the client is invoked with anything else than a Cozy
Note.

It's still a mystery why we haven't spotted this earlier.