gui: Hide dock icon on macOS by default (#2086)
Recent changes in Electron might have altered the behavior of
`Window.setVisibleOnAllWorkspaces()` effectively making the app icon in
the macOS dock show whenever we create a new window and call this
function (e.g. the Tray window which is hidden by default).

Besides, it seems that displaying an Electron app icon in the dock
prevents the OS from shutting down, displaying an error message to the
user inviting them to force-kill the application.

Since we don't need to show our windows on all workspaces (we only want
it displayed on the workspace it was opened from) we can stop using
this function and thus kill two birds with one stone by preventing
unexpected icons in the dock and leaving the OS shutdown correctly.