core/local/chokidar: Don't squash moves into moves (#2097)
On macOS, since local changes are buffered, when a document is moved
within a folder that is also moved, we end up with a move change for
which the old path (i.e. the path used to query PouchDB for the old
record) won't be valid anymore when it comes to merging this event.

e.g.:

  when doing those 2 moves:
  - `dir` → `my dir`
  - `dir/file` → `my dir/my file`

  the old path of the file move change will be `dir/file` while the
  saved path in the PouchDB record will be `my dir/file` once we've
  merged the directory move.

To handle those cases, we detect when a move is done inside another
move and update its old path to take into account its parent move.

However, the logic used to detect them, would also consider the move
of a document from outside the moved directory into it as a "move
inside a move" and try to "correct" its old path, leading to an
incorrect change.

e.g.:

  when doing those 2 moves:
  - `dir` → `my dir`
  - `file` → `my dir/file`

  the old path of the file move would be changed into `my dir` thus
  generating an invalid folder to file move.

We're now leaving those moves untouched as they should be.